Adds Fortran support for WY method & a regression test
Updates WY method to also support `tdy->ops` along with
the `SpatialFunction`. This changes allows one to reuse the
current Fortran interface.

Also, the F90 demo is renamed to reflect the change that it
now supports MPFA_O and WY.

Also, a new serial regression test is added for WY+F90.