Changed invocation of MatSetValuesLocal in 3D MPFAO TS.
Now we set all row values simultaneously. Regression tests pass.
I haven't profiled the performance to compare, but this should be
a bit faster.

I also moved variable declarations in the relevant function within
the scope of their usage. This is a modern way of declaring variables
that aids learning and debugging, because it makes clear how variables
are used instead of piling them at the top of a function. Unfortunately,
Fortran doesn't support it, so it might look a bit weird for some.