Got rid of set_snes and set_ts operations in favor of config_dm.
This operation configures the dycore's DM for its mode/discretization/etc.

This is just a sketch so far--the details of the operations needed to properly
configure a DM for a solver may evolve as we learn more.

I also renamed TDySetupNumericalMethods to TDySetup. This is about as
descriptive and is less typing. Hope nobody minds. :-)